+++ Volume/View Models/SettingsPageRow.swift
@@ -1,0 +1,55 @@
+//
+//  SettingsPageRow.swift
+//  Volume
+//
+//  Created by Cameron Russell on 4/15/21.
+//  Copyright © 2021 Cornell AppDev. All rights reserved.
+//
+
+import SwiftUI
+
+struct SettingsPageRow: View {
+    let page: Page
+    
+    private var row: some View {
+        HStack {
+            Image(page.imageName)
+                .resizable()
+                .renderingMode(.template)
+                .foregroundColor(Color.volume.lightGray)
+                .frame(width: 24, height: 24)
+                .padding()
+            Text(page.info)
+                .font(.helveticaRegular(size: 16))
+                .foregroundColor(.black)
+            Spacer()
+            Image("back-arrow")
+                .rotationEffect(Angle(degrees: 180))
+                .padding()
+        }
+        .padding([.leading, .trailing])
+    }
+    
+    var body: some View {
+        switch page.destination {
+        case .externalLink(let urlString):
+            if let url = URL(string: urlString) {
+                Link(destination: url) {
+                    row
+                }
+            }
+        case .internalView(let viewName):
+            if let destination = Settings.SettingView(rawValue: viewName) {
+                NavigationLink(destination: Settings.getView(for: destination)) {
+                    row
+                }
+            }
+        }
+    }
+}
+
+//struct SettingsPageRow_Previews: PreviewProvider {
+//    static var previews: some View {
+//        SettingsPageRow()
+//    }
+//}
