+++ Volume/Views/AboutUsView.swift
@@ -9,15 +9,8 @@
 import SwiftUI
 
 struct AboutUsView: View {
-    private var volume: VolumeInfo {
-        let filepath = Bundle.main.url(forResource: "VolumeInfo", withExtension: "json")!
-        let data = try! Data(contentsOf: filepath)
-        let result = try! JSONDecoder().decode(VolumeInfo.self, from: data)
-        return result
-    }
-    
     var body: some View {
-        let kvPairs = volume.subteams.sorted { $0.0.compare($1.0, options: .caseInsensitive) == .orderedAscending }
+        let kvPairs = Metadata.aboutUs.subteams.sorted { $0.0.compare($1.0, options: .caseInsensitive) == .orderedAscending }
         let subteams = kvPairs.map { $0.key }
         let subteamMembers = kvPairs.map { $0.value }
         
@@ -25,21 +18,34 @@ struct AboutUsView: View {
             ScrollView(showsIndicators: false) {
                 VStack(alignment: .leading) {
                     Header("Our Mission")
-                    Text(volume.messages.ourMission)
+                    Text(Metadata.aboutUs.messages.ourMission)
                         .font(.helveticaRegular(size: 16))
                         .lineSpacing(2)
                         .fixedSize(horizontal: false, vertical: true)
                         .padding([.bottom, .top])
                     
                     Header("The Team")
-                    Text(volume.messages.teamInfo)
+                    Text(Metadata.aboutUs.messages.teamInfo)
                         .font(.helveticaRegular(size: 16))
                         .lineSpacing(2)
                         .fixedSize(horizontal: false, vertical: true)
                         .padding([.bottom, .top])
 
                     ForEach(subteams.indices) { index in
-                        SubteamMembersView(subteam: subteams[index], members: subteamMembers[index])
+                        HStack {
+                            VStack {
+                                Text("📣")
+                                Spacer()
+                            }
+                            VStack(alignment: .leading, spacing: 3) {
+                                Text(subteams[index])
+                                    .font(.helveticaBold(size: 16))
+                                    .fixedSize(horizontal: false, vertical: true)
+                                Text(subteamMembers[index].joined(separator: ", "))
+                                    .fixedSize(horizontal: false, vertical: true)
+                            }
+                        }
+                        .padding([.bottom, .trailing], 15)
                     }
                 }
             }
@@ -48,38 +54,6 @@ struct AboutUsView: View {
         .padding(20)
         .background(Color.volume.backgroundGray)
     }
-}
-
-struct SubteamMembersView: View {
-    let subteam: String
-    let members: [String]
-    
-    var body: some View {
-        HStack {
-            VStack {
-                Text("📣")
-                Spacer()
-            }
-            VStack(alignment: .leading, spacing: 3) {
-                Text(subteam)
-                    .font(.helveticaBold(size: 16))
-                    .fixedSize(horizontal: false, vertical: true)
-                Text(members.joined(separator: ", "))
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-        .padding([.bottom, .trailing], 15)
-    }
-}
-
-struct VolumeInfo: Codable {
-    var messages: Messages
-    var subteams: [String: [String]]
-}
-
-struct Messages: Codable {
-    var ourMission: String
-    var teamInfo: String
 }
 
 //struct AboutUsView_Previews: PreviewProvider {
